restructured files, changed navbar select to button
