+++ .gitignore
@@ -9,6 +9,7 @@ conservation-layers.json
 conservation-authority-info.js
 webform-expressapp/views/mapiframe/conservation-layers.json
 /retired
+/.vscode
 # testing
 /coverage
