+++ src/components/graphapplication.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import DataContext from '../data';
+import DataContext from '../contexts/graphdata';
 import NavBar from './navbar';
 import CategoryList from './categorylist';
 import TabList from './tablist';
@@ -13,22 +13,22 @@ class GraphApplication extends Component {
             <LoadingOverlay active={this.context.loadingCA}
                             spinner
                             text="Loading Conservation Authorites....">
-            <div className="App" style={{marginBottom:"0px", marginTop:"0px"}}>
-                <div className="wrapper" style={{marginLeft:"40px", paddingTop:"15px", marginRight:"30px"}}> 
-                    <div className="headers text-left" style={{alignItems:"left"}}>
-                        <p><h1>Flood Hazard Mapping Analytics</h1></p>
-                        <p><h4>Filter By</h4></p>
-                        </div>
-                        <div className="row">
-                        <div className="col-lg-4 col-md-6 col-sm-12">
-                            <CategoryList></CategoryList>
-                        </div>
-                        <div className="col-lg-8 col-md-6 col-sm-12">
-                            <TabList></TabList>
+                <div className="App" style={{marginBottom:"0px", marginTop:"0px"}}>
+                    <div className="wrapper" style={{marginLeft:"40px", paddingTop:"15px", marginRight:"30px"}}> 
+                        <div className="headers text-left" style={{alignItems:"left"}}>
+                            <p><h1>Flood Hazard Mapping Analytics</h1></p>
+                            <p><h4>Filter By</h4></p>
+                            </div>
+                            <div className="row">
+                            <div className="col-lg-4 col-md-6 col-sm-12">
+                                <CategoryList></CategoryList>
+                            </div>
+                            <div className="col-lg-8 col-md-6 col-sm-12">
+                                <TabList></TabList>
+                            </div>
                         </div>
                     </div>
                 </div>
-            </div>
           </LoadingOverlay>
         );
     }
