+++ src/components/navbar.jsx
@@ -3,7 +3,7 @@ import Navbar from 'react-bootstrap/Navbar';
 import Nav from 'react-bootstrap/Nav'; 
 import { Form, FormControl, FormLabel, FormGroup } from 'react-bootstrap';
 
-const NavBar = () => {
+const NavBar = (props) => {
     return(
         <Navbar bg="primary" variant="dark" expand="lg">
             <Navbar.Brand href="#">Flood Mapping Metadata Portal</Navbar.Brand>
@@ -14,16 +14,7 @@ const NavBar = () => {
                     <Nav.Link href="/analytics" style={{color:"white"}}>View Data</Nav.Link>
                 </Nav>
                 <Nav className="mr-sm-2">
-                    <Form inline>
-                        <FormGroup controlId="modeselector">
-                            <FormLabel style={{color:"white", paddingRight: "8px"}}>Mode</FormLabel>
-                            <FormControl as="select" className="mr-sm-2">
-                                <option>Graphs</option>
-                                <option>Heatmaps</option>
-                            </FormControl>
-                        </FormGroup>
-                        
-                    </Form>
+                    <button type="button" className="btn btn-success" onClick={props.onButtonClick}>{props.children}</button>
                 </Nav>
             </Navbar.Collapse>
         </Navbar>
