+++ webform-expressapp/views/mapiframe/heatmapplugin.js
@@ -19,7 +19,7 @@ window.heatmap = {
         this.loadConservationLayers(); 
         //console.log($(".rv-legend-root"));
         window.addEventListener("message", (e) => {
-            if (e.data == "toggle conservation") {
+            if (e.data === "toggle conservation") {
                 const caLayer = api.layers.getLayersById("calayer")[0]; 
                 this.showCaLayer = !this.showCaLayer; 
                 caLayer.esriLayer.setVisibility(this.showCaLayer);
